fixup some tests
