+++ mismo/linker/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
+from mismo.linker._basic import EmptyLinkage as EmptyLinkage
+from mismo.linker._basic import EmptyLinker as EmptyLinker
+from mismo.linker._basic import FullLinkage as FullLinkage
+from mismo.linker._basic import FullLinker as FullLinker
+from mismo.linker._basic import UnnestLinker as UnnestLinker
+from mismo.linker._common import Linker as Linker
 from mismo.linker._id_linker import IDLinker as IDLinker
 from mismo.linker._join_linker import JoinLinker as JoinLinker
 from mismo.linker._key_linker import KeyLinker as KeyLinker
-from mismo.linker._linker import EmptyLinkage as EmptyLinkage
-from mismo.linker._linker import EmptyLinker as EmptyLinker
-from mismo.linker._linker import FullLinkage as FullLinkage
-from mismo.linker._linker import FullLinker as FullLinker
-from mismo.linker._linker import Linker as Linker
-from mismo.linker._linker import UnnestLinker as UnnestLinker
