+++ mismo/tests/test_resolve.py
@@ -59,13 +59,13 @@ def test_key_pair_resolvers_happy(condition, left, right):
     resolvers = _resolve.key_pair_resolvers(condition)
     assert len(resolvers) == 3
     r1, r2, r3 = resolvers
-    assert isinstance(r1, _resolve.IndividualDeferredResolver)
+    # assert isinstance(r1, _resolve.IndividualDeferredResolver)
     assert r1(left, right)[0].equals(left.username)
     assert r1(left, right)[1].equals(right.USERNAME)
-    assert isinstance(r2, _resolve.IndividualDeferredResolver)
+    # assert isinstance(r2, _resolve.IndividualDeferredResolver)
     assert r2(left, right)[0].equals(left.address)
     assert r2(left, right)[1].equals(right.address.upper())
-    assert isinstance(r3, _resolve.IndividualDeferredResolver)
+    # assert isinstance(r3, _resolve.IndividualDeferredResolver)
     assert r3(left, right)[0].equals(left.age)
     assert r3(left, right)[1].equals(right.age + 1)
 
@@ -75,7 +75,7 @@ def test_key_pair_resolvers_single(left, right):
     resolvers = _resolve.key_pair_resolvers(condition)
     assert len(resolvers) == 1
     (r1,) = resolvers
-    assert isinstance(r1, _resolve.IndividualDeferredResolver)
+    # assert isinstance(r1, _resolve.IndividualDeferredResolver)
     assert r1(left, right)[0].equals(left.username)
     assert r1(left, right)[1].equals(right.USERNAME)
 
@@ -85,7 +85,7 @@ def test_key_pair_resolvers_multiple_equals(left, right):
     resolvers = _resolve.key_pair_resolvers(condition)
     assert len(resolvers) == 1
     (r1,) = resolvers
-    assert isinstance(r1, _resolve.IndividualDeferredResolver)
+    # assert isinstance(r1, _resolve.IndividualDeferredResolver)
     assert r1(left, right)[0].equals(left.is_true)
     assert r1(left, right)[1].equals(right.USERNAME == "foo")
